Migrate repo search sample to be the test app
