--- samples/repo-search/android-views/build.gradle
@@ -1,17 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'org.jetbrains.kotlin.android'
-
-android {
-  namespace "com.example.redwood.reposearch.android.views"
-}
-
-dependencies {
-  implementation projects.samples.repoSearch.launcher
-  implementation projects.samples.repoSearch.presenterTreehouse
-  implementation projects.testApp.schema.widgetProtocol
-  implementation projects.redwoodLayoutView
-  implementation projects.redwoodLazylayoutView
-  implementation libs.androidx.appCompat
-  implementation libs.androidx.core
-  implementation libs.google.material
-}
