--- samples/repo-search/android-views/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-  <string name="app_name">Repo Search: Views</string>
-</resources>
