--- samples/repo-search/presenter-treehouse/build.gradle
@@ -1,34 +1,0 @@
-apply plugin: 'org.jetbrains.kotlin.multiplatform'
-apply plugin: 'org.jetbrains.kotlin.plugin.serialization'
-apply plugin: 'app.cash.zipline'
-
-kotlin {
-  jvm()
-
-  js {
-    // The name of the JS module which needs to be unique within the repo.
-    moduleName = 'repo-search-presenter-treehouse'
-    browser()
-    binaries.executable()
-  }
-
-  sourceSets {
-    commonMain {
-      dependencies {
-        api projects.redwoodTreehouse
-      }
-    }
-
-    jsMain {
-      dependencies {
-        api projects.redwoodTreehouseGuest
-        api projects.samples.repoSearch.presenter
-        implementation projects.testApp.schema.composeProtocol
-      }
-    }
-  }
-}
-
-zipline {
-  mainFunction = 'com.example.redwood.reposearch.treehouse.preparePresenters'
-}
