+++ test-app/android-views/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
