+++ test-app/android-views/src/main/res/mipmap-hdpi/ic_launcher.webp
