+++ test-app/android-views/src/main/res/mipmap-mdpi/ic_launcher.webp
