+++ test-app/android-views/src/main/res/mipmap-xhdpi/ic_launcher.webp
