+++ test-app/android-views/src/main/res/mipmap-xxhdpi/ic_launcher.webp
