+++ test-app/android-views/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
