+++ test-app/android-views/src/main/res/values/colors.xml
