+++ test-app/launcher/src/commonMain/kotlin/com/example/redwood/testing/launcher/TestAppSpec.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2023 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.example.redwood.testing.launcher
+
+import app.cash.redwood.treehouse.TreehouseApp
+import app.cash.zipline.Zipline
+import com.example.redwood.testing.treehouse.HostApi
+import com.example.redwood.testing.treehouse.TestAppPresenter
+import kotlinx.coroutines.flow.Flow
+
+class TestAppSpec(
+  override val manifestUrl: Flow<String>,
+  private val hostApi: HostApi,
+) : TreehouseApp.Spec<TestAppPresenter>() {
+  override val name = "test-app"
+
+  override fun bindServices(zipline: Zipline) {
+    zipline.bind<HostApi>("HostApi", hostApi)
+  }
+
+  override fun create(zipline: Zipline): TestAppPresenter {
+    return zipline.take<TestAppPresenter>("TestAppPresenter")
+  }
+}
