+++ test-app/presenter-treehouse/build.gradle
@@ -1,0 +1,34 @@
+apply plugin: 'org.jetbrains.kotlin.multiplatform'
+apply plugin: 'org.jetbrains.kotlin.plugin.serialization'
+apply plugin: 'app.cash.zipline'
+
+kotlin {
+  jvm()
+
+  js {
+    // The name of the JS module which needs to be unique within the repo.
+    moduleName = 'test-app-presenter-treehouse'
+    browser()
+    binaries.executable()
+  }
+
+  sourceSets {
+    commonMain {
+      dependencies {
+        api projects.redwoodTreehouse
+      }
+    }
+
+    jsMain {
+      dependencies {
+        api projects.redwoodTreehouseGuest
+        api projects.testApp.presenter
+        implementation projects.testApp.schema.composeProtocol
+      }
+    }
+  }
+}
+
+zipline {
+  mainFunction = 'com.example.redwood.testing.treehouse.preparePresenters'
+}
